Feature (ui, font, table): integrate color picker to cell properties
