+++ packages/ckeditor5-font/tests/manual/font-color-picker.js
@@ -9,6 +9,7 @@ import ClassicEditor from '@ckeditor/ckeditor5-editor-classic/src/classiceditor'
 import ArticlePluginSet from '@ckeditor/ckeditor5-core/tests/_utils/articlepluginset';
 import FontColor from '../../src/fontcolor';
 import FontBackgroundColor from '../../src/fontbackgroundcolor';
+import { TableProperties, TableCellProperties } from '@ckeditor/ckeditor5-table';
 
 ClassicEditor
 	.create( document.querySelector( '#editor-with-default-picker' ), {
@@ -16,7 +17,9 @@ ClassicEditor
 		plugins: [
 			ArticlePluginSet,
 			FontColor,
-			FontBackgroundColor
+			FontBackgroundColor,
+			TableProperties,
+			TableCellProperties
 		],
 		toolbar: [
 			'heading',
@@ -29,9 +32,13 @@ ClassicEditor
 			'bulletedList',
 			'numberedList',
 			'blockQuote',
+			'insertTable',
 			'undo',
 			'redo'
 		],
+		table: {
+			contentToolbar: [ 'tableColumn', 'tableRow', 'mergeTableCells', 'tableProperties', 'tableCellProperties' ]
+		},
 		fontColor: {
 			columns: 3
 		}
